2004-11-06  Paul Jenner  <psj.home@ntlworld.com>
	* java/awt/image/IndexColorModel.java (IndexColorModel): Fix
	constructor.


git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/branches/java-gui-branch@90165 138bc75d-0d04-0410-961f-82ee72b054a4

+++ libjava/java/awt/image/IndexColorModel.java
@@ -211,7 +211,7 @@ public class IndexColorModel extends ColorModel
    * @param trans the index of the transparent color
    * @param transferType DataBuffer.TYPE_BYTE or DataBuffer.TYPE_USHORT
    */
-  public IndexColorModel (int bits, int size, byte[] cmap, int start, 
+  public IndexColorModel (int bits, int size, int[] cmap, int start, 
                           boolean hasAlpha, int trans, int transferType)
   {
     super(bits * 4, // total bits, sRGB, four channels
